+++ terraform/modules/hello-nginx/variables.tf
@@ -1,0 +1,14 @@
+variable "deployment" {
+  type = bool
+  default = false
+}
+
+variable "service" {
+  type = bool
+  default = false
+}
+
+variable "ingress_https" {
+  type = bool
+  default = false
+}