+++ terraform/providers.tf
@@ -7,6 +7,9 @@ terraform {
     kubectl = {
       source = "gavinbunney/kubectl"
     }
+    helm = {
+      source = "hashicorp/helm"
+    }
   }
   required_version = "~> 1.5.0"
 }
@@ -15,4 +18,11 @@ terraform {
 provider "kubectl" {
   config_path    = "../ansible/static/kubeconfig"
   config_context = "kubernetes-admin@kubernetes"
+}
+
+provider "helm" {
+  kubernetes {
+    config_path    = "../ansible/static/kubeconfig"
+    config_context = "kubernetes-admin@kubernetes"
+  }
 }